deleted obsolete files
refactored unit tests to appropriate test files
