--- pkg/provisioner/client.go
@@ -1,17 +1,0 @@
-package provisioner
-
-/*import (
-	"context"
-
-	"k8s.io/apimachinery/pkg/runtime"
-	"github.com/yard-turkey/lib-bucket-provisioner/pkg/client/clientset/versioned"
-)*/
-
-// internalClient wraps related reconciler fields for convenient passing to helpers. Client is a Controller-runtime
-// DelegatingClient generated by the manager in manager.go. This is for optimization of API calls where in we make API
-// resource requests against the store and writes against the API server.
-// type internalClient struct {
-// 	client versioned.Interface
-// 	ctx    context.Context
-// 	scheme *runtime.Scheme
-// }
